requirements.txt
numpy
pandas
scikit-learn
nltk
spacy
sentence-transformers
pdfplumber
python-docx
fastapi
uvicorn
python-multipart
sentence-transformers
scikit-learn
pdfplumber
python-docx
python-dotenv
groq
